requirements.txt
click>=8.0.0
rich>=10.0.0
python-dotenv>=0.19.0
faiss-cpu>=1.7.0
numpy>=1.19.0
requests>=2.26.0
aiohttp>=3.8.0
duckduckgo-search>=3.0.0
azure-ai-inference
scipy>=1.7.0
